requirements.txt
numpy>=1.14.5
pandas>=0.24.2
torch>=0.4.1
tqdm
scikit-learn>=0.22.1
scipy>=1.1.*
tables
matplotlib
flowserv-core==0.2.0
